refactor: Restructure project
